feat : 코치님 피드백 반영 - 파사드에서 분리된 서비스를 호출하는 구조로 변경(테스트 아직 미반영
