+++ src/main/java/hhplus/concertreservationservice/application/concert/facade/ConcertPaymentFacade.java
@@ -3,6 +3,7 @@ package hhplus.concertreservationservice.application.concert.facade;
 import hhplus.concertreservationservice.application.concert.dto.ConcertCriteria;
 import hhplus.concertreservationservice.application.concert.dto.ConcertResult;
 import hhplus.concertreservationservice.domain.concert.dto.ConcertInfo;
+import hhplus.concertreservationservice.domain.concert.service.ConcertPaymentService;
 import hhplus.concertreservationservice.domain.concert.service.ConcertService;
 import hhplus.concertreservationservice.domain.queue.service.QueueService;
 import hhplus.concertreservationservice.global.exception.CustomGlobalException;
@@ -15,7 +16,7 @@ import org.springframework.stereotype.Component;
 public class ConcertPaymentFacade {
 
     private final QueueService queueService;
-    private final ConcertService concertService;
+    private final ConcertPaymentService concertPaymentService;
 
     public ConcertResult.Pay pay(ConcertCriteria.Pay criteria){
 
@@ -28,7 +29,7 @@ public class ConcertPaymentFacade {
         }
 
         // 예약내용 결제 (트랜잭션 적용. 위 서비스와 아래 서비스를 트랜잭션 분리 - 멘토링 반영)
-        ConcertInfo.Pay pay = concertService.payReservation(criteria.toCommand());
+        ConcertInfo.Pay pay = concertPaymentService.payReservation(criteria.toCommand());
 
         // 요구사항 : 결제가 완료되면 대기열 토큰을 만료시킵니다.
         queueService.expireToken(criteria.queueToken());
