+++ src/main/java/hhplus/concertreservationservice/application/concert/facade/ConcertReservationFacade.java
@@ -2,8 +2,12 @@ package hhplus.concertreservationservice.application.concert.facade;
 
 import hhplus.concertreservationservice.application.concert.dto.ConcertCriteria;
 import hhplus.concertreservationservice.application.concert.dto.ConcertResult;
+import hhplus.concertreservationservice.domain.concert.dto.ConcertInfo.ReserveSeat;
+import hhplus.concertreservationservice.domain.concert.service.ConcertPaymentService;
+import hhplus.concertreservationservice.domain.concert.service.ConcertReservationService;
 import hhplus.concertreservationservice.domain.concert.service.ConcertService;
 import hhplus.concertreservationservice.domain.queue.service.QueueService;
+import java.math.BigDecimal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class ConcertReservationFacade {
 
     private final QueueService queueService;
+    private final ConcertReservationService concertReservationService;
     private final ConcertService concertService;
 
 
@@ -21,14 +26,21 @@ public class ConcertReservationFacade {
         // 대기열 검증
         queueService.verifyQueue(criteria.queueToken());
 
-        // 예약 수행
-        return ConcertResult.ReserveSeat.fromInfo(concertService.reserveSeat(criteria.toCommand()));
+        // 좌석 관련 로직 서비스 호출
+        BigDecimal seatPrice = concertService.changeSeatStatusAndReturnPrice(criteria.concertSeatId());
+
+        // 예약 서비스 호출
+        ReserveSeat reserveSeatInfo = concertReservationService.reserveSeat(criteria.toCommand(seatPrice));
+
+        return ConcertResult.ReserveSeat.fromInfo(reserveSeatInfo);
 
     }
 
+
+    // 예약 만료 스케줄러.
     @Transactional
     public void expireReservationProcess() {
-        concertService.expireReservationProcess();
+        concertReservationService.expireReservationProcess();
     }
 
 }
