+++ src/main/java/hhplus/concertreservationservice/domain/concert/service/ConcertService.java
@@ -26,6 +26,7 @@ import hhplus.concertreservationservice.domain.user.repository.UserPointHistoryR
 import hhplus.concertreservationservice.domain.user.repository.UserRepository;
 import hhplus.concertreservationservice.global.exception.CustomGlobalException;
 import hhplus.concertreservationservice.global.exception.ErrorCode;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -38,15 +39,8 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ConcertService {
 
-    private static final Logger log = LoggerFactory.getLogger(ConcertService.class);
-
     private final ConcertScheduleRepository concertScheduleRepository;
     private final ConcertSeatRepository concertSeatRepository;
-    private final ConcertReservationRepository concertReservationRepository;
-    private final ConcertPaymentRepository concertPaymentRepository;
-    private final UserRepository userRepository;
-    private final QueueRepository queueRepository;
-    private final UserPointHistoryRepository userPointHistoryRepository;
 
     @Transactional(readOnly = true)
     public ConcertInfo.AvailableSchedules getAvailableSchedules(
@@ -57,7 +51,6 @@ public class ConcertService {
 
     }
 
-
     @Transactional(readOnly = true)
     public ConcertInfo.AvailableSeats getAvailableSeats(GetAvailableSeats command) {
         return AvailableSeats.fromEntityList(
@@ -65,10 +58,9 @@ public class ConcertService {
         );
     }
 
-
-    public ConcertInfo.ReserveSeat reserveSeat(ReserveSeat command) {
+    public BigDecimal changeSeatStatusAndReturnPrice(Long concertSeatId){
         // 좌석 조회( 낙관락 적용 )
-        ConcertSeat concertSeat = concertSeatRepository.findById(command.concertSeatId())
+        ConcertSeat concertSeat = concertSeatRepository.findById(concertSeatId)
             .orElseThrow(() -> new CustomGlobalException(ErrorCode.CONCERT_SEAT_NOT_FOUND));
 
         // 좌석 상태 변경
@@ -78,101 +70,9 @@ public class ConcertService {
             case SOLD -> throw new CustomGlobalException(ErrorCode.ALREADY_SOLD_SEAT);
         }
 
-        // 예약 생성
-        return ConcertInfo.ReserveSeat.builder()
-            .reservationId(concertReservationRepository.save(
-                    ConcertReservation.builder()
-                        .userId(command.userId())
-                        .concertSeatId(command.concertSeatId())
-                        .status(ReservationStatusType.RESERVED)
-                        .price(concertSeat.getPrice())
-                        .build())
-                .getId())
-            .build();
+        return concertSeat.getPrice();
     }
 
 
-    @Transactional
-    public ConcertInfo.Pay payReservation(Pay command) {
-        // 예약 조회
-        ConcertReservation reservation = concertReservationRepository.findById(
-                command.reservationId())
-            .orElseThrow(() -> new CustomGlobalException(ErrorCode.CONCERT_RESERVATION_NOT_FOUND));
 
-        // 예약 상태 검증 및 변경
-        reservation.confirmPayment();
-
-        // 유저 조회, 유저 잔액 검증-차감 (비관락 적용 + 유저의 잔액조회랑 다른 메서드 사용.)
-        User user = userRepository.findByIdForUsePoint(command.userId())
-            .orElseThrow(() -> new CustomGlobalException(ErrorCode.USER_NOT_FOUND));
-
-        if (user.getPoint().compareTo(reservation.getPrice()) > 0) {
-            // 변경감지. 포인트 사용
-            user.pointUse(reservation.getPrice());
-
-        } else {
-            throw new CustomGlobalException(ErrorCode.NOT_ENOUGH_BALANCE);
-        }
-
-        // 좌석 조회, 좌석 상태 변경
-        ConcertSeat concertSeat = concertSeatRepository.findById(reservation.getConcertSeatId())
-            .orElseThrow(() -> new CustomGlobalException(ErrorCode.CONCERT_SEAT_NOT_FOUND));
-
-        concertSeat.confirmSeatByPayment();
-
-        // 포인트 사용내역 저장
-        userPointHistoryRepository.save(UserPointHistory.builder()
-            .userId(user.getId())
-            .requestPoint(reservation.getPrice())
-            .type(UserPointHistoryType.PAY)
-            .build()
-        );
-
-        // 결제 생성 후 반환
-        return ConcertInfo.Pay.builder()
-            .paymentId(concertPaymentRepository.save(ConcertPayment.builder()
-                        .reservationId(reservation.getId())
-                        .price(reservation.getPrice())
-                        .status(PaymentStatusType.SUCCEED)
-                        .build()
-                    )
-                    .getId()
-            )
-            .build();
-    }
-
-    @Transactional
-    public void expireReservationProcess() {
-
-        // 만료된 예약 조회
-        List<ConcertReservation> expiredReservations = concertReservationRepository.findExpiredReservations(
-            ReservationStatusType.RESERVED,
-            LocalDateTime.now().minusMinutes(5));
-        log.info("expired reservation & seat size : {}, time : {}", expiredReservations.size(),
-            LocalDateTime.now());
-
-        // 만료된 예약 for루프 돌기
-        expiredReservations.forEach(reservation -> {
-            Queue queue = queueRepository.findByUserId(reservation.getUserId())
-                .orElseThrow(() -> new CustomGlobalException(ErrorCode.USER_NOT_FOUND));
-
-            // 대기열 제거
-            queueRepository.delete(queue);
-
-            //예약 취소 (RESERVED -> CANCELED)
-            reservation.cancelReservation();
-
-            //좌석 점유 해제
-            ConcertSeat reservedSeat = concertSeatRepository.findByIdAndStatus(
-                    reservation.getConcertSeatId(), SeatStatusType.RESERVED)
-                .orElseThrow(() -> new CustomGlobalException(ErrorCode.CONCERT_SEAT_NOT_FOUND));
-
-            // 좌석 상태 변경 (RESERVED -> EMPTY)
-            reservedSeat.cancelSeatByReservation();
-
-            log.info("expired reservation & seat id : {}, {}, time : {}", reservation.getId(),
-                reservedSeat.getId(), LocalDateTime.now());
-
-        });
-    }
 }
